Refactor code
for icons, adjust list function, id_col, and depended and in-depended vars
