+++ knime_extension/icons/icon/SpatialModel/GMendogErr.png
