+++ knime_extension/src/nodes/spatialmodels.py
@@ -18,55 +18,9 @@ __category = knext.category(
 __NODE_ICON_PATH = "icons/icon/SpatialModel/"
 
 
-def get_id_col_parameter(
-    label: str = "ID column",
-    description: str = """Select the column which contains for each observation in the input data a unique ID, it should be an integer column.
-    The IDs must match with the values of the 
-    [Spatial Weights node](https://hub.knime.com/center%20for%20geographic%20analysis%20at%20harvard%20university/extensions/sdl.harvard.features.geospatial/latest/org.knime.python3.nodes.extension.ExtensionNodeSetFactory$DynamicExtensionNodeFactory:4d710eae/)
-    ID column.
-    If 'none' is selected, the IDs will be automatically generated from 0 to the number of rows following the order of 
-    the first input table.
-    """,
-):
-    """
-    Returns the unique ID column. It should always keep the same as the ID column in the spatial weights matrix node.
-    The selected column should contain unique IDs for each observation in the input data.
-    """
-    return knext.ColumnParameter(
-        label=label,
-        description=description,
-        include_none_column=True,
-        column_filter=knut.is_long,
-        since_version="1.1.0",
-    )
-
-
 @knext.parameter_group("Advanced Settings", since_version="1.2.0")
 class AdvancedLsSetting:
     """Advanced Setting"""
-
-    # robust = knext.StringParameter(
-    #     "Robust",
-    #     """If ‘white’, then a White consistent estimator of the variance-covariance matrix is given.
-    #     If ‘hac’, then a HAC consistent estimator of the variance-covariance matrix is given. Default set to None.""",
-    #     enum=["white", "hac", "none"],
-    #     default_value="none",
-    #     is_advanced=True,
-    # )
-
-    # sig2n_k = knext.BoolParameter(
-    #     "Sig2n k",
-    #     """If True, then use n-k to estimate sigma^2. If False, use n. Default set to True.""",
-    #     default_value=True,
-    #     is_advanced=True,
-    # )
-
-    # spat_diag = knext.BoolParameter(
-    #     "Spat Diag",
-    #     """If True, then compute spatial diagnostics. Default set to False.""",
-    #     default_value=False,
-    #     is_advanced=True,
-    # )
 
     vm = knext.BoolParameter(
         "VM",
@@ -119,7 +73,7 @@ class Spatial2SLSModel:
     # input parameters
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
     dependent_variable = knext.ColumnParameter(
         "Dependent variable",
@@ -173,17 +127,7 @@ class Spatial2SLSModel:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
-
-        from libpysal.weights import W
-
-        w = W.from_adjlist(adjust_list)
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         y = gdf[self.dependent_variable].values
         x = gdf[self.independent_variables].values
 
@@ -350,19 +294,9 @@ class SpatialLagPanelModelwithFixedEffects:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.MultiColumnParameter(
-        "Dependent variables",
-        "The column containing the dependent variables to use for the calculation of Spatial Lag Panel Model with Fixed Effects.",
-        column_filter=knut.is_numeric,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of Spatial Lag Panel Model with Fixed Effects.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     advanced_settings = AdvancedLagErrorSetting()
 
@@ -374,18 +308,7 @@ class SpatialLagPanelModelwithFixedEffects:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
-
-        from libpysal.weights import W
-
-        w = W.from_adjlist(adjust_list)
-
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         x = gdf[self.independent_variables].values
         y = gdf[self.dependent_variable].values
 
@@ -498,19 +421,9 @@ class SpatialErrorPanelModelwithFixedEffects:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.MultiColumnParameter(
-        "Dependent variables",
-        "The column containing the dependent variables to use for the calculation of Spatial Error Panel Model with Fixed Effects.",
-        column_filter=knut.is_numeric,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of Spatial Error Panel Model with Fixed Effects.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     advanced_settings = AdvancedLagErrorSetting()
 
@@ -522,18 +435,7 @@ class SpatialErrorPanelModelwithFixedEffects:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
-
-        from libpysal.weights import W
-
-        w = W.from_adjlist(adjust_list)
-
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         x = gdf[self.independent_variables].values
         y = gdf[self.dependent_variable].values
 
@@ -681,17 +583,7 @@ class GeographicallyWeightedRegression:
 
     geo_col = knut.geo_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column containing the dependent variable to use for the calculation of Geographically Weighted Regression.",
-        column_filter=knut.is_numeric,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of Geographically Weighted Regression.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     fixed = knext.BoolParameter(
         "Fixed bandwidth",
@@ -900,11 +792,7 @@ class GeographicallyWeightedRegressionPredictor:
 
     geo_col = knut.geo_col_parameter()
 
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of Geographically Weighted Regression.",
-        column_filter=knut.is_numeric,
-    )
+    independent_variables = mut.get_dependent_and_independent_variables()[-1]
 
     def configure(self, configure_context, input_schema, input_binary_schema):
         self.geo_col = knut.column_exists_or_preset(
@@ -975,17 +863,7 @@ class MultiscaleGeographicallyWeightedRegression:
 
     geo_col = knut.geo_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column contains the dependent variable to use for the calculation of Multiscale Geographically Weighted Regression.",
-        column_filter=knut.is_numeric,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of Multiscale Geographically Weighted Regression.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     fixed = knext.BoolParameter(
         "Fixed bandwidth",
@@ -1130,73 +1008,6 @@ class MultiscaleGeographicallyWeightedRegression:
         # return knext.Table.from_pandas(gdf),model_string,knext.view_html(html)
 
 
-##############################################################################################################
-# Multiscale Geographically Weighted Regression Predictor Node
-# Notice that the prediction of MGWR is not yet implemented in the mgwr package
-##############################################################################################################
-
-# @knext.node(
-#     name="Multiscale Geographically Weighted Regression Predictor",
-#     node_type=knext.NodeType.PREDICTOR,
-#     category=__category,
-#     icon_path=__NODE_ICON_PATH + "SpatialWeight.png",
-#     after=""
-# )
-# @knext.input_table(
-#     name="Input Table",
-#     description="The input table containing the data to use for the prediction of Multiscale Geographically Weighted Regression.",
-# )
-# @knext.input_binary(
-#     name="Model",
-#     description="The model to use for the prediction of Multiscale Geographically Weighted Regression.",
-#     id="mgwr.gwr.MGWR",
-# )
-# @knext.output_table(
-#     name="Output Table",
-#     description="The output table containing the prediction of Multiscale Geographically Weighted Regression.",
-# )
-# class MGWRPredictorNode:
-#     """
-#     Multiscale Geographically Weighted Regression Predictor Node
-#     """
-
-#     geo_col = knext.ColumnParameter(
-#         "Geometry Column",
-#         "The column containing the geometry of the input table.",
-#         column_filter=knut.is_geo,
-#         include_row_key=False,
-#         include_none_column=False,
-#     )
-
-# # do we need this?
-#     independent_variables = knext.MultiColumnParameter(
-#         "Independent variables",
-#         "The columns containing the independent variables to use for the prediction of Multiscale Geographically Weighted Regression.",
-#         column_filter=knut.is_numeric
-#     )
-
-#     def configure(self, configure_context, input_schema,input_schema_2):
-#         self.geo_col = knut.column_exists_or_preset(configure_context, self.geo_col, input_schema, knut.is_geo)
-#         return None
-
-#     def execute(self, exec_context:knext.ExecutionContext, input_1, model):
-
-#             gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-#             #Prepare Georgia dataset inputs
-#             g_X = gdf[self.independent_variables].values
-#             u = gdf['geometry'].x
-#             v = gdf['geometry'].y
-#             g_coords = np.array(list(zip(u,v)))
-#             # g_X = (g_X - g_X.mean(axis=0)) / g_X.std(axis=0)
-
-#             #Calibrate MGWR model
-#             mgwr_model = pickle.loads(model)
-#             gdf.loc[:,'predy'] = mgwr_model.model.predict(g_coords, g_X).predictions
-#             gdf.reset_index(drop=True, inplace=True)
-
-#             return knext.Table.from_pandas(gdf)
-
-
 ############################################
 # spatial OLS node
 ############################################
@@ -1240,20 +1051,9 @@ class SpatialOLS:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column containing the dependent variable to use for the calculation of the spatial OLS model.",
-        column_filter=knut.is_numeric,
-        include_none_column=False,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of the spatial OLS model.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     Spatial_Diagnostics = knext.BoolParameter(
         "Spatial Diagnostics",
@@ -1308,17 +1108,7 @@ class SpatialOLS:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
-
-        from libpysal.weights import W
-
-        w = W.from_adjlist(adjust_list)
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         # Prepare Georgia dataset inputs
         X = gdf[self.independent_variables].values
         y = gdf[self.dependent_variable].values
@@ -1445,20 +1235,9 @@ class SpatialML_Lag:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column containing the dependent variable to use for the calculation of the spatial ML_Lag model.",
-        column_filter=knut.is_numeric,
-        include_none_column=False,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of the spatial ML_Lag model.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     advanced_settings = AdvancedLagErrorSetting()
 
@@ -1481,17 +1260,7 @@ class SpatialML_Lag:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
-
-        from libpysal.weights import W
-
-        w = W.from_adjlist(adjust_list)
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         # Prepare Georgia dataset inputs
         X = gdf[self.independent_variables].values
         y = gdf[self.dependent_variable].values
@@ -1603,20 +1372,9 @@ class SpatialML_Error:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column containing the dependent variable to use for the calculation of the spatial ML_Error model.",
-        column_filter=knut.is_numeric,
-        include_none_column=False,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of the spatial ML_Error model.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     advanced_settings = AdvancedLagErrorSetting()
 
@@ -1639,17 +1397,7 @@ class SpatialML_Error:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
-
-        from libpysal.weights import W
-
-        w = W.from_adjlist(adjust_list)
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         # Prepare Georgia dataset inputs
         X = gdf[self.independent_variables].values
         y = gdf[self.dependent_variable].values
@@ -1725,8 +1473,6 @@ SHOULD_NOT_CONTAIN_MISSING_VALUES = """
     [Missing Value](https://hub.knime.com/knime/extensions/org.knime.features.base/latest/org.knime.base.node.preproc.pmml.missingval.compute.MissingValueHandlerNodeFactory/) node to replace them."""
 
 # Root path for all node icons in this file
-__NODE_ICON_PATH_2 = "icons/icon/Geolab/"
-
 ############################################
 # spatial GM_Error node
 ############################################
@@ -1736,7 +1482,7 @@ __NODE_ICON_PATH_2 = "icons/icon/Geolab/"
     name="Spatial GM Error",
     node_type=knext.NodeType.LEARNER,
     category=__category,
-    icon_path=__NODE_ICON_PATH_2 + "GMErr.png",
+    icon_path=__NODE_ICON_PATH + "GMErr.png",
 )
 @knext.input_table(
     name="Input Table",
@@ -1772,21 +1518,9 @@ class SpatialGM_Error:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column containing the dependent variable to use for the calculation of the spatial GM Error model.",
-        column_filter=knut.is_numeric,
-        include_none_column=False,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of the spatial GM Error model.",
-        column_filter=knut.is_numeric,
-    )
-
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
     def configure(self, configure_context, input_schema, input_schema_2):
         self.geo_col = knut.column_exists_or_preset(
             configure_context, self.geo_col, input_schema, knut.is_geo
@@ -1796,19 +1530,10 @@ class SpatialGM_Error:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
-
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         import pandas as pd
         import spreg
-        from libpysal.weights import W
 
-        w = W.from_adjlist(adjust_list)
         # Prepare Georgia dataset inputs
         X = gdf[self.independent_variables].values
         y = gdf[self.dependent_variable].values
@@ -1872,7 +1597,7 @@ class SpatialGM_Error:
     name="Spatial GM Error Het",
     node_type=knext.NodeType.LEARNER,
     category=__category,
-    icon_path=__NODE_ICON_PATH_2 + "GMErrHet.png",
+    icon_path=__NODE_ICON_PATH + "GMErrHet.png",
 )
 @knext.input_table(
     name="Input Table",
@@ -1907,20 +1632,9 @@ class SpatialGM_Error_Het:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column containing the dependent variable to use for the calculation of the spatial GM Error Het model.",
-        column_filter=knut.is_numeric,
-        include_none_column=False,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of the spatial GM Error Het model.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     def configure(self, configure_context, input_schema, input_schema_2):
         self.geo_col = knut.column_exists_or_preset(
@@ -1931,18 +1645,11 @@ class SpatialGM_Error_Het:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
+
         import pandas as pd
         import spreg
-        from libpysal.weights import W
 
-        w = W.from_adjlist(adjust_list)
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         # Prepare Georgia dataset inputs
         X = gdf[self.independent_variables].values
         y = gdf[self.dependent_variable].values
@@ -2006,7 +1713,7 @@ class SpatialGM_Error_Het:
     name="Spatial GM Error Hom",
     node_type=knext.NodeType.LEARNER,
     category=__category,
-    icon_path=__NODE_ICON_PATH_2 + "GMErrHom.png",
+    icon_path=__NODE_ICON_PATH + "GMErrHom.png",
 )
 @knext.input_table(
     name="Input Table",
@@ -2042,20 +1749,9 @@ class SpatialGM_Error_Hom:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column containing the dependent variable to use for the calculation of the spatial GM Error Hom model.",
-        column_filter=knut.is_numeric,
-        include_none_column=False,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of the spatial GM Error Hom model.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     def configure(self, configure_context, input_schema, input_schema_2):
         self.geo_col = knut.column_exists_or_preset(
@@ -2066,19 +1762,10 @@ class SpatialGM_Error_Hom:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
-
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         import pandas as pd
         import spreg
-        from libpysal.weights import W
 
-        w = W.from_adjlist(adjust_list)
         # Prepare Georgia dataset inputs
         X = gdf[self.independent_variables].values
         y = gdf[self.dependent_variable].values
@@ -2142,7 +1829,7 @@ class SpatialGM_Error_Hom:
     name="Spatial GM Combo",
     node_type=knext.NodeType.LEARNER,
     category=__category,
-    icon_path=__NODE_ICON_PATH_2 + "GMcombo.png",
+    icon_path=__NODE_ICON_PATH + "GMcombo.png",
 )
 @knext.input_table(
     name="Input Table",
@@ -2178,20 +1865,9 @@ class SpatialGM_Combo:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column containing the dependent variable to use for the calculation of the spatial GM Combo model.",
-        column_filter=knut.is_numeric,
-        include_none_column=False,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of the spatial GM Combo model.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     def configure(self, configure_context, input_schema, input_schema_2):
         self.geo_col = knut.column_exists_or_preset(
@@ -2202,19 +1878,10 @@ class SpatialGM_Combo:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
-
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         import pandas as pd
         import spreg
-        from libpysal.weights import W
 
-        w = W.from_adjlist(adjust_list)
         # Prepare Georgia dataset inputs
         X = gdf[self.independent_variables].values
         y = gdf[self.dependent_variable].values
@@ -2278,7 +1945,7 @@ class SpatialGM_Combo:
     name="Spatial GM Combo Het",
     node_type=knext.NodeType.LEARNER,
     category=__category,
-    icon_path=__NODE_ICON_PATH_2 + "GMcomboHet.png",
+    icon_path=__NODE_ICON_PATH + "GMcomboHet.png",
 )
 @knext.input_table(
     name="Input Table",
@@ -2314,20 +1981,9 @@ class SpatialGM_Combo_Het:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column containing the dependent variable to use for the calculation of the spatial GM Combo Het model.",
-        column_filter=knut.is_numeric,
-        include_none_column=False,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of the spatial GM Combo Het model.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     def configure(self, configure_context, input_schema, input_schema_2):
         self.geo_col = knut.column_exists_or_preset(
@@ -2338,18 +1994,9 @@ class SpatialGM_Combo_Het:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         import pandas as pd
         import spreg
-        from libpysal.weights import W
-
-        w = W.from_adjlist(adjust_list)
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
 
         # Prepare Georgia dataset inputs
         X = gdf[self.independent_variables].values
@@ -2414,7 +2061,7 @@ class SpatialGM_Combo_Het:
     name="Spatial GM Combo Hom",
     node_type=knext.NodeType.LEARNER,
     category=__category,
-    icon_path=__NODE_ICON_PATH_2 + "GMcomboHom.png",
+    icon_path=__NODE_ICON_PATH + "GMcomboHom.png",
 )
 @knext.input_table(
     name="Input Table",
@@ -2449,20 +2096,9 @@ class SpatialGM_Combo_Hom:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column containing the dependent variable to use for the calculation of the spatial GM Combo Hom model.",
-        column_filter=knut.is_numeric,
-        include_none_column=False,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of the spatial GM Combo Hom model.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     def configure(self, configure_context, input_schema, input_schema_2):
         self.geo_col = knut.column_exists_or_preset(
@@ -2473,18 +2109,9 @@ class SpatialGM_Combo_Hom:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         import pandas as pd
         import spreg
-        from libpysal.weights import W
-
-        w = W.from_adjlist(adjust_list)
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
 
         # Prepare Georgia dataset inputs
         X = gdf[self.independent_variables].values
@@ -2547,7 +2174,7 @@ class SpatialGM_Combo_Hom:
     name="Spatial GM Endog Error",
     node_type=knext.NodeType.LEARNER,
     category=__category,
-    icon_path=__NODE_ICON_PATH_2 + "GMendogErr.png",
+    icon_path=__NODE_ICON_PATH + "GMendogErr.png",
 )
 @knext.input_table(
     name="Input Table",
@@ -2583,20 +2210,9 @@ class SpatialGM_Endog_Error:
 
     geo_col = knut.geo_col_parameter()
 
-    id_col = get_id_col_parameter()
+    id_col = mut.get_id_col_parameter()
 
-    dependent_variable = knext.ColumnParameter(
-        "Dependent variable",
-        "The column containing the dependent variable to use for the calculation of the spatial GM Endog Error model.",
-        column_filter=knut.is_numeric,
-        include_none_column=False,
-    )
-
-    independent_variables = knext.MultiColumnParameter(
-        "Independent variables",
-        "The columns containing the independent variables to use for the calculation of the spatial GM Endog Error model.",
-        column_filter=knut.is_numeric,
-    )
+    (dependent_variable, independent_variables) = mut.get_dependent_and_independent_variables()
 
     yend = knext.ColumnParameter(
         "Endogenous variable",
@@ -2621,19 +2237,10 @@ class SpatialGM_Endog_Error:
 
     def execute(self, exec_context: knext.ExecutionContext, input_1, input_2):
         gdf = gp.GeoDataFrame(input_1.to_pandas(), geometry=self.geo_col)
-        adjust_list = input_2.to_pandas()
-
-        if "none" not in str(self.id_col).lower():
-            gdf.index = range(len(gdf))
-            id_map = dict(zip(gdf[self.id_col], gdf.index))
-            adjust_list["focal"] = adjust_list["focal"].map(id_map)
-            adjust_list["neighbor"] = adjust_list["neighbor"].map(id_map)
-
+        w, gdf = mut.get_w_from_adjust_list(input_2, gdf, self.id_col)
         import pandas as pd
         import spreg
-        from libpysal.weights import W
 
-        w = W.from_adjlist(adjust_list)
         # Prepare Georgia dataset inputs
         X = gdf[self.independent_variables].values
         y = gdf[self.dependent_variable].values
